+++ need2give/lib/screens/user/item.dart
@@ -4,7 +4,7 @@ import 'package:need2give/constants/global.dart';
 import 'package:need2give/models/item.dart';
 import 'package:need2give/screens/donation_center/update_item.dart';
 import 'package:need2give/screens/donation_center/bottom_bar.dart';
-import 'package:need2give/screens/user/donation_center/profile_screen.dart';
+import 'package:need2give/screens/user/donation_profile.dart';
 import 'package:need2give/services/item_service.dart';
 import 'package:need2give/widgets/textfield.dart';
